+++ projects/spring-cloud-petclinic-service/src/main/resources/messages/messages.properties
@@ -2,7 +2,7 @@
 # Canned Messages
 #
 
-welcome=Welcome To This *** NEW *** Place!
+welcome=Welcome To This Place!
 required=is required
 notFound=has not been found
 duplicate=is already in use
