+++ projects/spring-cloud-petclinic-service/src/main/resources/templates/welcome.html
@@ -10,10 +10,14 @@
     <p>Application Version: <span th:text="${applicationVersion}"> </span></p>
 
     <br>
-    <p>Current Date-Time: <span th:text="${currentDate}"> </span></p>
+    <p>  HostName: <span th:text="${hostName}"> </span></p>
 
     <br>
-    <p>HostName: <span th:text="${hostName}"> </span></p>
+    <p>IP-Address: <span th:text="${ipAddress}"> </span></p>
+
+    <br>
+    <p>Current Date-Time: <span th:text="${currentDate}"> </span></p>
+
 
     <br>
     <br>
